move digitizer helpers from base.py to digitizers.py
